Mutually Recursive Modules
Shifted some major stuffs around to allow for requiring of multiple modules. The drawback being that when loading the plugins module needs to be created before any requiring can happen. 

I'm unsure if this is how module loading is going to stay, I may opt for a less asynchronous, mutually recursive mess.